refactor: reorganize components by usage pattern
Implements issue #39 - restructure components into desktop/, mobile/, and shared/ directories.

## New Directory Structure

**Desktop Components** (`src/components/desktop/`)
- layout.ts (renamed from desktop.ts)
- header.ts (renamed from desktopHeader.ts)
- benchmark.ts, leaderboard.ts, progressBar.ts, ruleset.ts
- simulation.ts, summary.ts, theme.ts, dataMode.ts
- patternInspector.ts, zoomSlider.ts

**Mobile Components** (`src/components/mobile/`)
- layout.ts (renamed from mobile.ts)
- header.ts (renamed from mobileHeader.ts)
- statsOverlay.ts, buttonContainer.ts, roundButton.ts, starButton.ts

**Shared Components** (`src/components/shared/`)
- simulationInfo.ts, stats.ts (already in shared/)

## Import Path Updates

All imports updated to reflect new structure:
- Core module imports: `../` → `../../` (e.g., `../../api/save`)
- Shared components: `./shared/` → `../shared/` (from mobile/desktop)
- Local components: Updated to use new names (desktopHeader → header)
- Main entry point: Updated to use new paths (desktop/layout.ts, mobile/layout.ts)

## Benefits

- **Clear separation**: Desktop, mobile, and shared code are visually distinct
- **Better organization**: Related components grouped by usage pattern
- **Easier navigation**: Developers can quickly find platform-specific code
- **Improved maintainability**: Changes to mobile/desktop are isolated
- **Type safety**: All imports resolved correctly, 0 TypeScript errors

## Testing

- ✅ TypeScript compilation passes (0 errors)
- ✅ Build succeeds
- ✅ All imports resolve correctly
- ✅ Git history preserved via `git mv`

Resolves #39
