+++ src/components/desktop/patternInspector.ts
