+++ src/components/desktop/progressBar.ts
