+++ src/components/desktop/simulation.ts
