+++ src/components/desktop/zoomSlider.ts
