+++ src/components/mobile/starButton.ts
